+++ data_simulation.h
@@ -28,13 +28,13 @@ extern "C" {
 #endif
 
 void data_simulation(double *wavenumbers, double *spec, double channel_bandwidth_hz, int time_acc, unsigned int num_channels,
-                     unsigned int num_baselines, double sigma, unsigned long int n_gal, double g1, double g2,
+                     unsigned int num_baselines, double sigma, unsigned int n_gal, double g1, double g2,
                      double *ge1, double *ge2, double *gflux, double *gscale, double *l, double *m,
-                     double *SNR_vis, unsigned long int num_coords, double *uu_metres, double *vv_metres,
+                     double *SNR_vis, unsigned int num_coords, double *uu_metres, double *vv_metres,
                      double *ww_metres, complexd *visGal, complexd* visData);
 
 void sky_model(double *wavenumbers, double *spec, double channel_bandwidth_hz, int time_acc, unsigned int num_channels,
-                     unsigned long int n_gal, double *gflux, double *gscale, double *l, double *m, unsigned long int num_coords,
+                     unsigned int n_gal, double *gflux, double *gscale, double *l, double *m, unsigned int num_coords,
                      double *uu_metres, double *vv_metres, double *ww_metres, complexd *visGal, complexd* visMod);
 
 #ifdef __cplusplus
